refactor personalize part

+++ src/offline/news/step-funcs/build.sh
@@ -78,13 +78,16 @@ do
     --parameter-overrides ${PARAMETER_OVERRIDES} \
     --capabilities CAPABILITY_NAMED_IAM
 
-    echo "finish deploy"
 
-    StackStatus=$($AWS_CMD  cloudformation  describe-stacks --region ${REGION} --stack-name ${STACK_NAME} --output table | grep StackStatus)
-    echo ${StackStatus} |  egrep "(CREATE_COMPLETE)|(UPDATE_COMPLETE)" > /dev/null
+     StackStatus=$($AWS_CMD  cloudformation  describe-stacks --region ${REGION} --stack-name ${STACK_NAME} --output table | grep StackStatus)
+     echo ${StackStatus} |  egrep "(CREATE_COMPLETE)|(UPDATE_COMPLETE)" > /dev/null
 
-    rm tmp_*.yaml > /dev/null 2>&1 || true
+     if [[ $? -ne 0 ]]; then
+         echo "error!!!  ${StackStatus}"
+         exit 1
+     fi
 
+    rm tmp_*.yaml > /dev/null 2>&1  || true
 
 done
